requirements.txt
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
beautifulsoup4==4.11.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
cssselect2==0.7.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.7
django-adminactions==2.0.0
django-easy-pdf==0.1.1
django-filter==23.2
django-grappelli==3.0.4
django-import-export==3.1.0
django-single-session==0.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
html5lib==1.1
idna==3.4
lxml==4.9.2
Markdown==3.4.3
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.1
oscrypto==1.3.0
Pillow==9.5.0
psycopg2==2.9.5
pycparser==2.21
pyHanko==0.17.2
pyhanko-certvalidator==0.20.1
PyJWT==2.7.0
pypdf==3.8.1
pypng==0.20220715.0
python-bidi==0.4.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.4.2
reportlab==3.6.12
requests==2.28.2
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
svglib==1.5.1
tablib==3.3.0
tinycss2==1.2.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
uritools==4.0.1
urllib3==1.26.15
webencodings==0.5.1
xhtml2pdf==0.2.10
xlrd==2.0.1
XlsxWriter==3.0.9
xlwt==1.3.0
